pyproject.toml
[project]
name = "nonebot-plugin-bmclapi"
dynamic = ["version"]
description = "Get OpenBMCLAPI status in chat apps."
authors = [{ name = "student_2333", email = "lgc2333@126.com" } , { name = "XieXiLin", email = "support@xiexilin.com" }]
dependencies = [
    "nonebot2>=2.0.0",
    "pydantic<2,>=1.10.4",
    "nonebot-plugin-alconna>=0.33.8",
    "humanize>=4.9.0",
]
requires-python = ">=3.8,<4.0"
readme = "README.md"
license = { text = "MIT" }

[project.urls]
homepage = "https://github.com/lgc-NB2Dev/nonebot-plugin-bmclapi"

[tool.pdm.build]
includes = []

[tool.pdm.version]
source = "file"
path = "nonebot_plugin_bmclapi/__init__.py"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
